refactor: break up test/prevalence.test.ts
